pyproject.toml
[project]
name = "ableton-live-mcp-server"
version = "0.1.0"
description = "Ableton Live MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.4.1",
    "pydantic-ai>=0.0.49",
    "python-dotenv>=1.0.1",
    "python-osc>=1.9.3",
    "ruff>=0.11.2",
    "typer>=0.15.2",
]
